requirements.txt
aiohttp
gdown
python-docx
EbookLib
beautifulsoup4
pdfminer.six
connexion[swagger-ui]
aiohttp_jinja2
